Time: 1 ms (100%), Space: 62.5 MB (10.07%) - LeetHub

+++ 3689-maximum-total-subarray-value-i/3689-maximum-total-subarray-value-i.java
@@ -1,6 +1,6 @@
 class Solution {
     public long maxTotalValue(int[] nums, int k) {
-        long max = 0, min = Integer.MAX_VALUE;
+        int max = 0, min = Integer.MAX_VALUE;
 
         for (int num: nums) {
             max = Math.max(num, max);
